Add files via upload
One of my personal favorite school assignments.
This program creates a ball, with a mouse click, and the ball bounces up and down on the screen.

Assignment details:
Create a bouncing ball simulation in which a single ball (thread) bounces up and down in a vertical line. The ball should bounce off the bottom and top of the enclosing frame.

Once you have that working, modify your program to allow the user to add additional balls to the simulation by pressing the space bar or clicking the mouse.